Extend ngRepeat to work properly with ngForm
ngRepeat doesn't work as expected with form rows dynamically output
using ngForm. This directive requires that you name your ng-form "form".
It then makes your dynamically output rows properties of your scope's
main form controller in the format of propertyName_$id where $id is the
id of the ng-repeat iteration. This allows for regular form controller
validation and etc to be used on a dynamic form.